fix: move components to category directories and fix font-family
Address PR review feedback:
- Move components to category subdirectories (navigation, menus,
  content, overlays, inputs) matching phase-5 structure
- Fix font-family to use var(--rr-font-family-body) in all components
- Update component-map.json, index.ts, and story paths
