+++ src/components/index.ts
@@ -86,22 +86,24 @@ export { RRListItemDragHandle } from './lists/list-item-drag-handle/rr-list-item
 export { RRListItemDragHandleCell } from './lists/list-item-drag-handle-cell/rr-list-item-drag-handle-cell.ts';
 export { RRTimelineTrackCell } from './lists/timeline-track-cell/rr-timeline-track-cell.ts';
 
-// Navigation components
-export { RRTabBarItem } from './tab-bar/rr-tab-bar-item.ts';
-export { RRTabBar } from './tab-bar/rr-tab-bar.ts';
-export { RRDocumentTabBarItem } from './document-tab-bar-item/rr-document-tab-bar-item.ts';
-export { RRPagination } from './pagination/rr-pagination.ts';
-export { RRStandaloneMenuItem } from './menu-item/rr-standalone-menu-item.ts';
+// Navigation components (continued)
+export { RRTabBarItem } from './navigation/tab-bar/rr-tab-bar-item.ts';
+export { RRTabBar } from './navigation/tab-bar/rr-tab-bar.ts';
+export { RRDocumentTabBarItem } from './navigation/document-tab-bar-item/rr-document-tab-bar-item.ts';
+export { RRPagination } from './navigation/pagination/rr-pagination.ts';
+
+// Menu components
+export { RRStandaloneMenuItem } from './menus/menu-item/rr-standalone-menu-item.ts';
 
 // Input components (continued)
-export { RRPasswordField } from './password-field/rr-password-field.ts';
+export { RRPasswordField } from './inputs/password-field/rr-password-field.ts';
 
-// Typography components
-export { RRRichTextHeading } from './rich-text-heading/rr-rich-text-heading.ts';
+// Content components
+export { RRRichTextHeading } from './content/rich-text-heading/rr-rich-text-heading.ts';
 
 // Overlay components
-export { RRTooltipArrow } from './tooltip/rr-tooltip-arrow.ts';
-export { RRTooltip } from './tooltip/rr-tooltip.ts';
-export { RRDialog } from './dialog/rr-dialog.ts';
+export { RRTooltipArrow } from './overlays/tooltip/rr-tooltip-arrow.ts';
+export { RRTooltip } from './overlays/tooltip/rr-tooltip.ts';
+export { RRDialog } from './overlays/dialog/rr-dialog.ts';
 
 // Auto-register happens on import of individual component files
