+++ src/components/inputs/password-field/rr-password-field.ts
@@ -29,7 +29,7 @@ export class RRPasswordField extends LitElement {
   static override styles = css`
     :host {
       display: block;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
