+++ src/components/navigation/document-tab-bar-item/rr-document-tab-bar-item.ts
@@ -33,7 +33,7 @@ export class RRDocumentTabBarItem extends LitElement {
   static override styles = css`
     :host {
       display: block;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
