+++ src/components/navigation/pagination/rr-pagination.ts
@@ -23,7 +23,7 @@ export class RRPagination extends LitElement {
   static override styles = css`
     :host {
       display: inline-block;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
