+++ src/components/navigation/tab-bar/rr-tab-bar-item.stories.js
@@ -11,7 +11,7 @@ export default {
       url: 'https://www.figma.com/design/5DyHMXUNVxbgH7ZjhQxPZe/RR-Components?node-id=1366-43604',
     },
     componentSource: {
-      file: 'src/components/tab-bar/rr-tab-bar-item.ts',
+      file: 'src/components/navigation/tab-bar/rr-tab-bar-item.ts',
       repository: 'https://github.com/regelrecht/design-system',
     },
     status: { type: 'stable' },
