+++ src/components/navigation/tab-bar/rr-tab-bar.ts
@@ -20,7 +20,7 @@ export class RRTabBar extends LitElement {
   static override styles = css`
     :host {
       display: inline-block;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
