+++ src/components/overlays/dialog/rr-dialog.ts
@@ -29,7 +29,7 @@ export class RRDialog extends LitElement {
   static override styles = css`
     :host {
       display: contents;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
