+++ src/components/overlays/tooltip/rr-tooltip.ts
@@ -24,7 +24,7 @@ export class RRTooltip extends LitElement {
   static override styles = css`
     :host {
       display: inline-flex;
-      font-family: var(--rr-font-family-sans, 'RijksSansVF', system-ui, sans-serif);
+      font-family: var(--rr-font-family-body);
     }
 
     :host([hidden]) {
